+++ packages/schemas/package.json
@@ -5,9 +5,9 @@
   "license": "MIT",
   "private": true,
   "main": "src/index.ts",
+  "types": "src/index.ts",
   "typings": "src/index.ts",
   "files": [
-    "dist",
     "src"
   ],
   "scripts": {
