+++ packages/server/package.json
@@ -18,7 +18,6 @@
   "dependencies": {
     "@fastify/cors": "^8.2.1",
     "@onix/utils": "workspace:*",
-    "alchemy-sdk": "^2.6.2",
     "axios": "^1.3.4",
     "bignumber.js": "^9.1.1",
     "fastify": "^4.15.0"
