+++ packages/server/src/app.ts
@@ -1,21 +1,20 @@
 import Fastify from 'fastify';
 import cors from '@fastify/cors';
-import type { AddressDetails } from '@onix/schemas';
-import { assets } from '@onix/utils';
-import { BigNumber } from 'bignumber.js';
-import { Etherscan } from './etherscan';
-import { CoinMarketCap } from './coinmarketcap';
+import { Client } from './client';
 import { addressDetailsParams } from './schemas';
-import { take, toBaseUnit } from './utils';
 
 const router = Fastify({
-  logger: process.env.NODE_ENV === 'dev',
+  logger: process.env.NODE_ENV === 'development',
 });
 
 router.register(cors, { origin: true });
 
-const etherscan = new Etherscan();
-const coinmarketcap = new CoinMarketCap();
+const client = new Client({
+  apiKeys: {
+    etherscan: process.env.ETHERSCAN_API_KEY,
+    coinmarketcap: process.env.COINMARKETCAP_API_KEY,
+  },
+});
 
 router
   .get('/_health', () => {
@@ -23,50 +22,7 @@ router
   })
   .get('/address/:address', async (req) => {
     const { address } = addressDetailsParams.parse(req.params);
-    const etherBalance = await etherscan.getEtherBalance(address);
-    const etherPrice = await etherscan.getEtherPrice();
-
-    // Limit to 3 assets for now due to rate limits
-    const tokens = take(assets, 3);
-    const tokensWithBalances = await Promise.all(
-      tokens.map(async (token) => {
-        const balance = await etherscan.getERC20Balance(address, token.address);
-        return {
-          ...token,
-          balance: {
-            token: new BigNumber(balance).div(new BigNumber(10).pow(token.decimals)).toFixed(4),
-            // To be set once token prices have been fetched
-            usd: '0',
-          },
-        };
-      }),
-    );
-
-    const tokenSymbols = tokens.map((token) => token.symbol);
-    const tokenPrices = await coinmarketcap.getTokenPrices(tokenSymbols);
-
-    tokensWithBalances.forEach((asset, index) => {
-      asset.balance.usd = new BigNumber(asset.balance.token)
-        .times(tokenPrices[index].quote.USD.price)
-        .toFixed(4);
-    });
-
-    const totalBalance = tokensWithBalances.reduce(
-      (acc, asset) => acc.plus(new BigNumber(asset.balance.usd)),
-      new BigNumber(0),
-    );
-
-    const result: AddressDetails = {
-      address,
-      etherBalance: {
-        token: toBaseUnit(etherBalance, 18).toFixed(4),
-        usd: toBaseUnit(etherBalance, 18).times(etherPrice).toFixed(4),
-      },
-      totalBalance: totalBalance.toFixed(4),
-      assets: tokensWithBalances,
-    };
-
-    return result;
+    return client.getAddressDetails(address);
   });
 
 const start = async () => {
